refactor(frontend): extract chats tags perspectives and health api clients
