+++ frontend/src/modules/chats/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./chats.api";
