+++ frontend/src/modules/health/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./infrastructure";
