+++ frontend/src/modules/health/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./health.api";
