+++ frontend/src/modules/perspectives/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./perspectives.api";
