+++ frontend/src/modules/tags/index.ts
@@ -1,1 +1,2 @@
 export * from "./domain";
+export * from "./infrastructure";
