+++ frontend/src/modules/tags/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./tags.api";
